refactor: CLI moved to separate package
